pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "leap"
version = "1.0.0"
description = "Live Experiments for Active Pedagogy"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "sqlalchemy>=2.0",
    "duckdb-engine>=0.9.0",
    "typer>=0.9.0",
    "requests>=2.28.0",
    "pyyaml>=6.0",
    "itsdangerous>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx>=0.24.0",
]

[project.scripts]
leap = "leap.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["leap*"]
